+++ versionhash.go
@@ -1,0 +1,2 @@
+package main
+var GitCommitHash = "0e5f587a077810d058202b76a127651a02bd4382"