+++ versiontag.go
@@ -1,0 +1,2 @@
+package main
+var GitTag = "v0.10.6"